test: split secrets runtime integration coverage
